Fixed race condition where the KILL signal was sent many times. Now the devserver sends a kill signal, sets the server to null, and restarts the server once the original server's exit callback is called.

+++ devserver.js
@@ -7,21 +7,31 @@ var child_js_file = process.ARGV[2];
 var current_dir = __filename.split('/');
 current_dir = current_dir.slice(0, current_dir.length-1).join('/');
 
-var start_server = function(){
+function start_server () {
     server = spawn('node', [child_js_file]);
-    function pass_along(data){
-        if (data !== null) sys.print(data);
-    }
+    server.addListener('exit', onServerExit);
     server.addListener('output', pass_along);
     server.addListener('error', pass_along);
     sys.puts('server started');
-};
+}
 
-var restart_server = function(){
-    sys.puts('change discovered, restarting server');
-    server.kill('SIGHUP');
-    setTimeout(start_server, 500);
-};
+function restart_server () {
+    // make sure we only send a kill signal once
+    if(server != null) { 
+        sys.puts('Change discovered, restarting server');
+        server.kill('SIGHUP');
+        server = null;
+    }
+}
+
+function onServerExit(exitCode) {
+    sys.puts('Server process exited with code ' + exitCode);
+    start_server();
+}
+
+function pass_along(data){
+    if (data !== null) sys.print(data);
+}
 
 function parse_file_list (dir, files) {
